fix for deprecation of implicit dtype conversion

+++ staircase/core/stairs.py
@@ -33,9 +33,9 @@ from staircase.util._decorators import Appender
 
 def _make_deltas_from_vals(init_val, vals: pd.DataFrame) -> pd.Series:
     if not np.isnan(vals).any():
-        result = pd.Series(np.diff((np.append([init_val], vals.values))))
+        result = pd.Series(np.diff((np.append([init_val], vals.values))), dtype="float64")
     else:
-        temp = pd.Series(np.append([init_val], vals.values))
+        temp = pd.Series(np.append([init_val], vals.values), dtype="float64")
         result = pd.Series.add(
             temp[pd.notnull(temp)].diff(), temp[pd.isnull(temp)], fill_value=0
         )[1:]
